requirements.txt
requests==2.28.2
schedule==1.1.0
yadisk==1.2.14
revChatGPT~=2.3.2
asyncio~=3.4.3
Flask~=2.2.3
instagrapi~=1.16.41
pytest~=7.2.1
python-telegram-bot~=20.1
emoji~=2.2.0
pillow~=9.4.0